+++ Classes/Record/RecordRegistry.php
@@ -5,10 +5,8 @@ declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\Record;
 
 use TYPO3\CMS\Core\Cache\CacheManager;
-use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Core\SingletonInterface;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use YoastSeoForTypo3\YoastSeo\Exception\UnsupportedVersionException;
 
 class RecordRegistry implements SingletonInterface
 {
@@ -24,12 +22,6 @@ class RecordRegistry implements SingletonInterface
 
     public function addRecord(string $tableName): Record
     {
-        if (GeneralUtility::makeInstance(Typo3Version::class)->getMajorVersion() < 10) {
-            throw new UnsupportedVersionException(
-                'The record feature is only available for TYPO3 10 and higher',
-                1659438424
-            );
-        }
         return $this->records[$tableName]
             ?? ($this->records[$tableName] = GeneralUtility::makeInstance(Record::class)->setTableName($tableName));
     }