+++ ext_tables.php
@@ -1,40 +1,26 @@
 <?php
 
-if (TYPO3_MODE === 'BE') {
-    $offset = 0;
-    foreach ($GLOBALS['TBE_MODULES'] as $key => $_) {
-        if ($key === 'web') {
-            $GLOBALS['TBE_MODULES'] = array_slice($GLOBALS['TBE_MODULES'], 0, ($offset + 1), true) +
-                ['yoast' => ''] +
-                array_slice($GLOBALS['TBE_MODULES'], $offset + 1, count($GLOBALS['TBE_MODULES']) - 1, true);
-        }
-        $offset++;
-    }
+defined('TYPO3') || die;
 
-    $GLOBALS['TBE_MODULES']['_configuration']['yoast'] = [
-        'iconIdentifier' => 'module-yoast',
-        'labels' => 'LLL:EXT:yoast_seo/Resources/Private/Language/BackendModule.xlf',
-        'name' => 'yoast'
-    ];
-
-    $registerExtension = 'YoastSeo';
-    $moduleController = \YoastSeoForTypo3\YoastSeo\Controller\ModuleController::class;
-    $overviewController = \YoastSeoForTypo3\YoastSeo\Controller\OverviewController::class;
-    if (\TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(\TYPO3\CMS\Core\Information\Typo3Version::class)
-            ->getMajorVersion() < 10) {
-        $registerExtension = 'YoastSeoForTypo3.yoast_seo';
-        $moduleController = 'Module';
-        $overviewController = 'Overview';
-    }
+(static function () {
+    \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addModule(
+        'yoast',
+        '',
+        'after:web',
+        null,
+        [
+            'iconIdentifier' => 'module-yoast',
+            'labels' => 'LLL:EXT:yoast_seo/Resources/Private/Language/BackendModule.xlf',
+            'name' => 'yoast'
+        ]
+    );
 
     \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerModule(
-        $registerExtension,
+        'YoastSeo',
         'yoast',
         'dashboard',
         '',
-        [
-            $moduleController => 'dashboard',
-        ],
+        [\YoastSeoForTypo3\YoastSeo\Controller\ModuleController::class => 'dashboard'],
         [
             'access' => 'user,group',
             'icon' => 'EXT:yoast_seo/Resources/Public/Images/Yoast-module-dashboard.svg',
@@ -43,13 +29,11 @@ if (TYPO3_MODE === 'BE') {
     );
 
     \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerModule(
-        $registerExtension,
+        'YoastSeo',
         'yoast',
         'overview',
         '',
-        [
-            $overviewController => 'list',
-        ],
+        [\YoastSeoForTypo3\YoastSeo\Controller\OverviewController::class => 'list'],
         [
             'access' => 'user,group',
             'icon' => 'EXT:yoast_seo/Resources/Public/Images/Yoast-module-overview.svg',
@@ -58,13 +42,11 @@ if (TYPO3_MODE === 'BE') {
     );
 
     \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerModule(
-        $registerExtension,
+        'YoastSeo',
         'yoast',
         'premium',
         '',
-        [
-            $moduleController => 'premium',
-        ],
+        [\YoastSeoForTypo3\YoastSeo\Controller\ModuleController::class => 'premium'],
         [
             'access' => 'user,group',
             'icon' => 'EXT:yoast_seo/Resources/Public/Images/Yoast-module-premium.svg',
@@ -77,9 +59,10 @@ if (TYPO3_MODE === 'BE') {
         'label' => 'LLL:EXT:yoast_seo/Resources/Private/Language/BackendModule.xlf:usersettings.hideYoastInPageModule',
         'type' => 'check'
     ];
-    $GLOBALS['TYPO3_USER_SETTINGS']['showitem'] .= ',
-            --div--;LLL:EXT:yoast_seo/Resources/Private/Language/BackendModule.xlf:usersettings.title,hideYoastInPageModule';
+    \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addFieldsToUserSettings(
+        '--div--;LLL:EXT:yoast_seo/Resources/Private/Language/BackendModule.xlf:usersettings.title,hideYoastInPageModule'
+    );
 
     $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_pagerenderer.php']['render-preProcess'][]
-        = \YoastSeoForTypo3\YoastSeo\Hooks\YoastConfigInlineJs::class . '->renderJsonConfig';
-}
+        = \YoastSeoForTypo3\YoastSeo\Hooks\BackendYoastConfig::class . '->renderConfig';
+})();